refactor(times): split TrainrunsectionHelper.getTravelTime() into 2 functions
`TrainrunsectionHelper.getTravelTime()` computes the new travel time of a
trainrunSection, either by using the `travelTimeFactor` and the current
travel time of the section, or by using the total trainrun travel time and
the sum of the new travel times computed up to that point.

This logic can be split into 2 independent functions, which will be used
depending on whether the considered section is the last one to compute or not.

Signed-off-by: Clara Ni <clara.ni@outlook.fr>