import: set node position before calling replaceIntermediateStopWithNode()
replaceIntermediateStopWithNode() recomputes a bunch of transition
position stuff. At import time, tt was working with incorrect node
positions, routing everything to the (0, 0) position. This causes
transition positions to be mis-rendered.

Instead, set the node position at initialization, before calling
replaceIntermediateStopWithNode(). That way the function can work
with the final node position. As a bonus, the "new node" event is
fired with the correct node position.

Signed-off-by: Simon Ser <contact@emersion.fr>
Fixes: 435f1c66b8ed ("import: handle importing old files with numberOfStops")