Handle nodesOrdered in TrainrunsectionHelper.getLeftAndRightTimes()
When nodesOrdered is empty, the closest stop nodes are picked. When
nodesOrdered isn't empty, these nodes should be used instead.

nodesOrdered isn't an ideal mechanism and it would be nice to
replace it with something else, but let's leave this for later.

PerlenketteSectionComponent.getTravelTime() needs to be adapted
because it relied on getTimeStructure().travelTime being the
cummulative travel time. This is no longer the case since the
perlenkette can edit the non-cummulative travel time only.

Signed-off-by: Simon Ser <contact@emersion.fr>