refactor: stop copying TrainrunSectionNodePair objects
Thanks to the previous commit, these objects are now immutable, so
no need to copy them.

Object.assign() creates a new object with the same properties.
However, the new object doesn't inherit the prototype, so looses
the class. As a result, class methods are undefined on the new
object.

TrainrunSectionNodePair doesn't have methods so this wasn't an
issue until now, but we'll be introducing a new method soon.

Signed-off-by: Simon Ser <contact@emersion.fr>