chore: introduce DirectedTrainrunSectionProxy
This will be used to de-duplicate a lot of logic where the
source-to-target and target-to-source sides are symmetrical. There
are numerous instances of this pattern in NGE, e.g. time
propagation.

"Head" and "tail" were picked because they haven't an overloaded
meaning in NGE yet and are standard graph theory parlance.

Signed-off-by: Simon Ser <contact@emersion.fr>